chore: add missing option cookie to URL screenshot
