+++ src/chromium/interfaces/common.types.ts
@@ -2,6 +2,16 @@ import { PathLikeOrReadStream } from '../../common';
 
 export type EmulatedMediaType = 'screen' | 'print';
 
+export type Cookie = {
+    name: string;
+    value: string;
+    domain: string;
+    path?: string;
+    secure?: boolean;
+    httpOnly?: boolean;
+    sameSite?: 'Strict' | 'Lax' | 'None';
+};
+
 export type ChromiumOptions = {
     header?: PathLikeOrReadStream;
     footer?: PathLikeOrReadStream;
@@ -12,4 +22,5 @@ export type ChromiumOptions = {
     failOnHttpStatusCodes?: number[]; // Return a 409 Conflict response if the HTTP status code is in the list (default [499,599])
     failOnConsoleExceptions?: boolean; // Return a 409 Conflict response if there are exceptions in the Chromium console (default false)
     skipNetworkIdleEvent?: boolean; // Do not wait for Chromium network to be idle (default false)
+    cookies?: Cookie[]; // Cookies to be written.
 };
